refactor: 매직 넘버 제거, 메서드 추출, 유틸클래스 생성자 제한
