+++ src/main/java/lotto/util/ShuffleNumber.java
@@ -1,10 +1,13 @@
-package lotto.domain;
+package lotto.util;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 public class ShuffleNumber {
+
+    private ShuffleNumber() {
+    }
 
     public static List<Integer> shuffle(int start, int end) {
         List<Integer> numbers = new ArrayList<>();
